Init
This is the initial commit of the pre-pre-pre-pre alpha. In this state the server runs colyseus and a websocket router for communication with the client. The client connects and you can control your dot on the screen. As other players they can control their dots. We see the dots move and update for all other logged in dots.

This will progress into a modular framework with a parody main story line that aims to rip on Vanilla World of Warcraft. With a tenured story and thousands of imtes and quests to make fun of the story mode content will be extensive.

The goal here is to allow players to run their own servers where they can contribute their own dungeons, areas, classes, crafts, crafted items, different level caps. The game is the engine and the server-client relationship.

We will add things to this as well such as farming and player housing so you can actually farm. The spell, and item system can be manipulated with some very easy json and scripting. Server operators can choose a classic "parody" server where players level up to level 60 and battle the usual suspects in this new 2D way or do they change it completely and instead of mages and paladins they have spell blades and reapers.

This will progress as devlopment continues. I'd like to go 3D but my skillset is limited. I can program a 2D game and slap assets together and make a story and quests and do level design. Adding 3D collision is beyond my capabilities without having more folks on board.

This is Classicraft where you can craft even the Classes.